refactor server client
